Battery gauge fix (#52)
* Fix battery reporting in status.json, and adjust scaling for bettery level representation

* remove comment verbosity

* change battery_value_svc to return float

+++ components/services/battery.c
@@ -43,7 +43,7 @@ static struct {
 /****************************************************************************************
  * 
  */
-int battery_value_svc(void) {
+float battery_value_svc(void) {
 	return battery.avg;
  }
